requirements.txt
# Kafka
kafka-python==2.0.2

# AWS SDK
boto3==1.34.34

# Data processing
pandas==2.2.0
numpy==1.26.3

# PySpark (for local development)
pyspark==3.5.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
